Fix projectile position

+++ RadDeg.cpp
@@ -1,12 +1,12 @@
 #include "stdafx.h"
 #include "RadDeg.h"
 
-float pi = 3.1415f; // You could be more precise!
+float pi = 3.1415; // You could be more precise!
 
-double degToRad(float degAngle) {
+double degToRad(double degAngle) {
 	return degAngle / 180 * pi;
 }
 
-double radToDeg(float radAngle) {
+double radToDeg(double radAngle) {
 	return radAngle / pi * 180;
 }
